requirements.txt
flask
groq
python-dotenv
gunicorn
uuid
datetime